refactor: rename the repo to be runtime agnostic (#189)
Renamed reference to the repo in preparation for the name change. It's
worth mentioning that the repo name should be changed before this PR is
merged. Also, the existing git tags will keep the old naming convention.

While this isn't a breaking change, it will require consumers to update
the package name in order to stay current.

Signed-off-by: Michael Beemer <beeme1mr@users.noreply.github.com>

Signed-off-by: Michael Beemer <beeme1mr@users.noreply.github.com>