fix: Fixes for #461
Fixes:
We remove all handlers from the old provider when a new one is set - but other clients could still be listening. We need to remove only some.

When we run handlers immediately in the situation a client is already READY, we don't wrap the handler in a try/catch and run it async.

Signed-off-by: Lukas Reining <lukas.reining@codecentric.de>