feat: react sdk groundwork (#546)
This PR:

- adds all the necessary groundwork for a new react module in the
monorepo (packaging, release, CI, etc)
- adds a _very_ limited implementation based on work by @moredip , see
example in
[README](https://github.com/open-feature/js-sdk/blob/a9104c1ca2a2fecd75b02e1d804a7bb984dd6d2d/packages/react/README.md)

Things to note:

- I've chosen NOT to bundle in the web-sdk directly into the react-sdk
(which I could have done). Instead, the web-sdk is a peer dependency.
Reasoning:
- NPM 16+ automatically downloads peer deps if missing, and otherwise
uses the available peer if compatible
- Providers "don't care" whether they are used by the web-sdk, or the
react-sdk, however, they all need a peerDependency on the web-sdk. If we
bundled the web-sdk into the react-sdk automatically, it would mean that
a consumer only using the react-sdk + some-web-provider would get a
"useless" copy of the web-sdk downloaded automatically. For that reason,
it's better to have the web-sdk as a peer, with the react SDK purely
building on to of that dep, not duplicating/embedding a bunch of its
code.
- the web-sdk objects are re-exported from the react-sdk, so consumers
only need to import from there.

relates to: https://github.com/open-feature/js-sdk/pull/546

:mega: I want to make clear that this is not a final API, and I'd love
to continue discussion in the linked issue on how this could work. We
can merge this PR (and subsequent ones) without releasing so that people
can experiment locally with various APIs, via
[yalc](https://github.com/wclr/yalc) or `npm link`.

---------

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>