feat: add provider compatibility check (#537)
## This PR

- add provider compatibility check

### Related Issues
<!-- add here the GitHub issue that this PR resolves if applicable -->

Fixes #535 

### Notes

This implementation throws if you set a provider that's incompatible
with its intended use. This could easily be changed to a warning if
that's the preferred approach. I think it's a good idea to fail early
but I would be interested in others thoughts on this.

This is a non-breaking change. Providers can optionally specify their
intended use and the SDK will perform a runtime check during provider
registration.

### Follow up

- Add hook compatibility checks.
- Update readme to include examples.
- Update existing providers.

---------

Signed-off-by: Michael Beemer <michael.beemer@dynatrace.com>