chore: expose transaction context propagation to the server sdk only (#590)
## This PR

- removes transaction context propagation from web SDK
- refactored the transaction context propagation logic to live under the
server SDK

### Notes

Transaction context propagation only makes sense on the server and was
simply ignored by the client. This removes the methods from the client
sdk to avoid confusion.

This is a **breaking change** in the web SDK. However, it's sub 1.0 and
this feature effectively had no effect if it was used. Currently, this
is not marked as a breaking change to avoid releasing a major version.
@toddbaert do you know if it's possible to mark the server sdk changes
as a chore and the web sdk changes as a breaking fix?

Signed-off-by: Michael Beemer <beeme1mr@users.noreply.github.com>

---------

Signed-off-by: Michael Beemer <beeme1mr@users.noreply.github.com>