chore: fix React SDK README.md to use the resolved value instead of the resolution details
The documentation conditionally renders something if the returned value from the hook is truthy, but this implies that it is simply checking for the entirety of the `EvaluationDetails<T>` returned from `useFeatureFlag` which seems off since we will always get an object back according to the typings. It feel like it should be checking against the `.value` to conditionally render the elements.

Signed-off-by: Juan Bernal <juanxwtf@gmail.com>