feat: add evaluation context management to the web SDK (#704)
## This PR

- adds the ability to manage named context in the web SDK

### Related Issues

Fixes #685 

### Notes

This change requires additions to the spec before it can be merged. I'll
mark this PR as a draft until the spec has been updated.

I also noticed that we currently use the term "named clients" to
describe the scoped providers. I believe this terminology is confusing
and made writing the JSDocs difficult because the client name shouldn't
be important when setting context. I think we should consider using the
term "Provider Namespace". In my opinion, this more accurately describes
the behavior and could be introduced in a non-breaking way.

---------

Signed-off-by: Michael Beemer <beeme1mr@users.noreply.github.com>