feat: add PROVIDER_CONTEXT_CHANGED event (web-sdk only) (#731)
This PR:

- adds `PROVIDER_CONTEXT_CHANGED` events, which, in the static paradigm,
can be used to inform the SDK that the flags should be re-evaluated
(important for UI repaints in React, for instance (note this event is
only available in the web-sdk)
- runs the associated `PROVIDER_CONTEXT_CHANGED` handlers if the
provider's context handler function ran successfully or `PROVIDER_ERROR`
handlers otherwise.
- adds associated tests

A decent amount of this is just typing magic to reduce duplicated code
while making the new event only available in the web-sdk.

See: [associated spec
change](https://github.com/open-feature/spec/pull/200)

Fixes: https://github.com/open-feature/js-sdk/issues/729

---------

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>