feat: suspense support, client scoping, context-sensitive re-rendering (#698)
Adds a few react features:

- `<Suspense />` support: components using feature flags will trigger
suspense for easy loaders/spinners
- Ability to specify name for provider scope: `<OpenFeatureProvider
name="my-provider">`

> [!IMPORTANT]  
> Please see
[here](https://github.com/open-feature/react-test-app/pull/2) for the
latest changes to the demo application using these features (note that
we won't be able to merge these demos until this is released).

> [!IMPORTANT]  
> Also note I've added no tests, which is certainly not my MO. I will
add them to this PR once there's an agreement on this behavior and
implementation.

gif from the demo app:


![demo](https://github.com/open-feature/js-sdk/assets/25272906/73007cff-0d0c-44e8-a34c-b99412556206)

:warning: I want to add another feature here to support re-rendering on
context changes. That requires
[this](https://github.com/open-feature/js-sdk/pull/731) to be merged.

---------

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>
Co-authored-by: Michael Beemer <beeme1mr@users.noreply.github.com>