fix(types): conflicts with peer types (#852)
With our recent changes in bundling, there's a new problem with the
framework-specific SDKs (react + nestjs). These have peers, and the
external function was bundling those peers as well, which can cause tsc
errors due to conflicts. This change prevents us from bundling any peers
(which is what we want, so we avoid conflicts with the consumers
installed peers).

This has no runtime impact, and there was no changes to core, web, or
server SDK dists. See image below for the difference in the react SDK
(notice peers are now IMPORTED):


![image](https://github.com/open-feature/js-sdk/assets/25272906/0bdd16d6-d1c0-4118-9bf4-93659c164090)

---------

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>