feat: Angular SDK (#997)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR
<!-- add the description of the PR here -->

Adds an Angular SDK as discussed in #976.
It mainly provides directives that conditionally render based on flag
value and provider state.

I did not inject any OpenFeature service or anything like that, as I
think the directives will be the way to access feature flags in the
templates, and the OpenFeature global API can be used directly in
services. But I am not 100% sure on that one.

Happy to receive feedback and opinions :)

A small test app can be found here:
https://github.com/open-feature/angular-test-app
This can be used with [yalc](https://github.com/wclr/yalc).

---------

Signed-off-by: Lukas Reining <lukas.reining@codecentric.de>
Co-authored-by: Todd Baert <todd.baert@dynatrace.com>