fix(angular): fix race condition on initialization (#1052)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR
<!-- add the description of the PR here -->

Fixes a race condition when the directive is first rendered.
Initially the initialization code in the constructor before the
`@Input`s were bound to the variables.
This meant that in some cases providers were called with `undefined` as
flag key, which most providers accept, but for flipt-web the engine
paniced due to the value being `unit` instead of a string.

---------

Signed-off-by: Lukas Reining <lukas.reining@codecentric.de>