chore: enable release-as=patch for release-please to trigger patch release for all packages (#1259)
Enabling` "release-as": "patch"` in the `release-please-config.json`
config per-package will trigger patch releases for each package,
allowing us to test NPM OIDC connections with a patch release.

This will be removed after a successful deployment.

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Signed-off-by: Jonathan Norris <jonathan@taplytics.com>