chore: test OIDC Changes with Angular Release (#1260)
Making a minimal change to trigger an Angular SDK patch release (0.0.18)
to validate the recent OIDC npm auth updates.

Chose Angular as it's the lowest-downloaded package, reducing risk for
this validation test.

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

Signed-off-by: Jonathan Norris <jonathan@taplytics.com>