chore: Test OIDC NPM Publishing Without Token (#1262)
Testing Angular SDK patch release after removing NPM_TOKEN from GitHub
secrets to validate that OIDC authentication alone works for npm
publishing.

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->

---------

Signed-off-by: Jonathan Norris <jonathan@taplytics.com>