fix: multi-provider hook context management
Fixing an issue with the MultiProvider where hook contexts and hints
were being lost due to copies of the context data being created in the
OpenFeature sdk evaluation.

Since key evaluation of Maps using objects is done by reference, the
lookup of the context during evaluation was failing, leading to errors.

Signed-off-by: Mike Kitzman <mdkitzman@gmail.com>