fix: multi-provider hook context management (#1282)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR
Fixing an issue with the MultiProvider where hook contexts and hints
were being lost due to copies of the context data being created in the
OpenFeature sdk evaluation.

Since key evaluation of Maps using objects is done by reference, the
lookup of the context during evaluation was failing, leading to errors.

- adds this new feature

### Related Issues

Fixes #1268

### Notes

### Follow-up Tasks

### How to test

---------

Signed-off-by: Mike Kitzman <mdkitzman@gmail.com>