feat: allow method to be passed into setContext hook (#1301)
## This PR

Allows for `setContext` returned by the `useContextMutator` hook to be
passed a method that returns the new context, with the current
(previous) context passed into it. This aligns the signature of
`setContext` to the setter of a regular `useState` hook, making it
easier to conditionally update the context without needing to keep track
of the domain separately from the React context.

---------

Signed-off-by: MattIPv4 <me@mattcowley.co.uk>