fix: use actual previous context to noop useContextMutator setContext (#1305)
## This PR

Updates the no-op logic in useContextMutator's setContext method to
compare against the actual previous context as returned by OpenFeature
itself, rather than just the previous context that that specific use of
useContextMutator had in its ref
(https://github.com/open-feature/js-sdk/pull/1301#discussion_r25658360790).

Building on #1301, this should allow for setContext to be called with a
method, and for that method to return the context it was provided, to
guarantee a no-op always.

Signed-off-by: MattIPv4 <me@mattcowley.co.uk>
Co-authored-by: Lukas Reining <lukas.reining@codecentric.de>