fix: remove explicit return type for React `OpenFeatureProvider` (#1313)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR

This PR removes the explicit `JSX.Element` return type from the type
signature of the React SDK `OpenFeatureProvider`.
`JSX.Element` is narrow for React 19+ as types have widened to include
Portals etc. & this was throwing TS compilation errors


### Related Issues

Fixes #1307

### Notes
As this is purely a type change I didn't feel it necessary to add
additional tests, happy to investigate further if required though

Signed-off-by: Sam <sam@andomain.co.uk>